requirements.txt
Django
djangorestframework
psycopg2

pandas
numpy
# django_redis
# django_celery_beat
openpyxl>=3.0,<4.0


Django>=3.0,<4.0
psycopg2-binary>=2.8
celery
redis
